[CRUD] 문의사항 답변 CRUD 구현

+++ src/pages/contact/contactboarddetail/ContactBoardDetail.tsx
@@ -12,54 +12,67 @@ function ContactBoardDetail() {
     title: '',
     date: '',
     content: '',
-    answer: ''
+    answer: '',
+    answer_id: null
   });
   const [answer, setAnswer] = useState('');
-  const [isAdmin, setIsAdmin] = useState(false); // 초기값을 null로 설정하여 초기 상태를 구분
+  const [isAdmin, setIsAdmin] = useState(false);
+  const [showAnswer, setShowAnswer] = useState(false);
+  const [isEditing, setIsEditing] = useState(false);
 
   useEffect(() => {
     const fetchDataAndCheckAdmin = async () => {
       try {
         const token = sessionStorage.getItem('isLogin');
-        //console.log('Token:', token); // 토큰이 올바르게 저장되었는지 확인
+        if (!token) {
+          console.error('No token found');
+          return;
+        }
 
         // 관리자 여부 확인
         const checkResponse = await axios.get('/check', {
           headers: {
-            Authorization: token // 토큰이 없을 때를 대비해 기본값 설정
+            Authorization: token
           }
         });
-        //console.log('Admin Check Response:', checkResponse.data); // 응답 데이터를 확인
         setIsAdmin(checkResponse.data);
 
         // 데이터 불러오기
         const response = await axios.get(`/question/${id}`, {
           headers: {
-            Authorization: token // 토큰이 없을 때를 대비해 기본값 설정
+            Authorization: token
           }
         });
-        //console.log('Question Data:', response.data); // 응답 데이터를 확인
         setDetailBoardData(response.data);
-        setAnswer(response.data.answer);
+
+        // 답변 데이터 불러오기
+        const answerResponse = await axios.get(`/question/${id}/answer/read`, {
+          headers: {
+            Authorization: token
+          }
+        });
+        setAnswer(answerResponse.data.answer);
+        setDetailBoardData((prevData: any) => ({ ...prevData, answer_id: answerResponse.data.id }));
       } catch (error) {
         console.error('Failed to fetch data or check admin status:', error);
-        setIsAdmin(false); // 에러 발생 시 관리자가 아니라고 설정
       }
     };
 
     fetchDataAndCheckAdmin();
-  }, [id, isAdmin]);
-
-  // useEffect(() => {
-  //   console.log('isAdmin:', isAdmin); // isAdmin 상태가 제대로 업데이트되었는지 확인
-  // }, [isAdmin]);
+  }, [id]);
 
   const handleDelete = async () => {
     if (window.confirm('삭제하겠습니까?')) {
       try {
+        const token = sessionStorage.getItem('isLogin');
+        if (!token) {
+          console.error('No token found');
+          return;
+        }
+
         await axios.delete(`/question/${id}/delete`, {
           headers: {
-            Authorization: sessionStorage.getItem('isLogin')
+            Authorization: token
           }
         });
         alert('삭제되었습니다.');
@@ -76,40 +89,147 @@ function ContactBoardDetail() {
   };
 
   const handleAnswerSubmit = async () => {
-    console.log('Submitting answer:', answer); // 제출될 답변 확인
+    console.log('Submitting answer:', answer);
     if (answer.trim() === '') {
       alert('답변을 입력해주세요.');
       return;
     }
     try {
-      await axios.post(`/answer/${id}`, { answer }, {
+      const token = sessionStorage.getItem('isLogin');
+      if (!token) {
+        console.error('No token found');
+        return;
+      }
+
+      const endpoint = isEditing
+        ? `/question/${id}/answer/update/${detailBoardData.answer_id}`
+        : `/question/${id}/answer/create`;
+
+      await axios.post(endpoint, { content: answer }, {
         headers: {
-          Authorization: sessionStorage.getItem('isLogin')
+          Authorization: token
         }
       });
       alert('답변이 제출되었습니다.');
-      setDetailBoardData((prevData: any) => ({ ...prevData, answer }));
-      setAnswer('');
+
+      // 답변 데이터 다시 불러오기
+      const answerResponse = await axios.get(`/question/${id}/answer/read`, {
+        headers: {
+          Authorization: token
+        }
+      });
+      setAnswer(answerResponse.data.answer);
+      setDetailBoardData((prevData: any) => ({ ...prevData, answer: answerResponse.data.answer, answer_id: answerResponse.data.id }));
+      setIsEditing(false);
+      setShowAnswer(false); // 수정 후 답변 입력란 닫기
+      window.location.reload(); // 페이지 새로고침
     } catch (error) {
       console.error('Failed to submit answer:', error);
       alert('답변 제출에 실패했습니다. 다시 시도해 주세요.');
     }
   };
 
+  const handleAnswerEdit = () => {
+    setIsEditing(true);
+    setShowAnswer(true);
+  };
+
+  const handleAnswerUpdate = async () => {
+    console.log('Updating answer:', answer);
+    if (answer.trim() === '') {
+      alert('답변을 입력해주세요.');
+      return;
+    }
+    try {
+      const token = sessionStorage.getItem('isLogin');
+      if (!token) {
+        console.error('No token found');
+        return;
+      }
+
+      await axios.put(`/question/${id}/answer/update/${detailBoardData.answer_id}`, { content: answer }, {
+        headers: {
+          Authorization: token
+        }
+      });
+      alert('답변이 수정되었습니다.');
+
+      // 수정된 답변 데이터 다시 불러오기
+      const answerResponse = await axios.get(`/question/${id}/answer/read`, {
+        headers: {
+          Authorization: token
+        }
+      });
+      setAnswer(answerResponse.data.answer);
+      setDetailBoardData((prevData: any) => ({ ...prevData, answer: answerResponse.data.answer, answer_id: answerResponse.data.id }));
+      setIsEditing(false);
+      setShowAnswer(false); // 수정 후 답변 입력란 닫기
+      window.location.reload(); // 페이지 새로고침
+    } catch (error) {
+      console.error('Failed to update answer:', error);
+      alert('답변 수정에 실패했습니다. 다시 시도해 주세요.');
+    }
+  };
+
+  const handleAnswerDelete = async () => {
+    if (window.confirm('답변을 삭제하겠습니까?')) {
+      try {
+        const token = sessionStorage.getItem('isLogin');
+        if (!token) {
+          console.error('No token found');
+          return;
+        }
+
+        await axios.delete(`/question/${id}/answer/delete/${detailBoardData.answer_id}`, {
+          headers: {
+            Authorization: token
+          }
+        });
+        alert('답변이 삭제되었습니다.');
+        setAnswer('');
+        setDetailBoardData((prevData: any) => ({ ...prevData, answer: '', answer_id: null }));
+        window.location.reload(); // 페이지 새로고침
+      } catch (error) {
+        console.error('Failed to delete answer:', error);
+        alert('답변 삭제에 실패했습니다. 다시 시도해 주세요.');
+      }
+    }
+  };
+
+  const toggleAnswerSection = () => {
+    setShowAnswer(!showAnswer);
+  };
+
   const renderAnswerSection = () => {
-    if (isAdmin === null) {
-      return <C.NoAnswerText>관리자 확인 중...</C.NoAnswerText>;
-    } else if (isAdmin) {
-      return (
-        <>
-          <C.AnswerTextArea value={answer} onChange={handleAnswerChange} placeholder="답변을 입력해주세요." />
-          <C.SubmitButton onClick={handleAnswerSubmit}>제출</C.SubmitButton>
-        </>
-      );
-    } else if (detailBoardData.answer) {
-      return detailBoardData.answer;
+    if (isAdmin) {
+      if (detailBoardData.answer && !isEditing) {
+        return (
+          <C.AnswerContainer>
+            {detailBoardData.answer}
+            <C.BottomButtonWrapper>
+              <C.ActionButton onClick={handleAnswerEdit}>수정</C.ActionButton>
+              <C.ActionButton onClick={handleAnswerDelete}>삭제</C.ActionButton>
+            </C.BottomButtonWrapper>
+          </C.AnswerContainer>
+        );
+      } else {
+        return (
+          <C.AnswerSection>
+            <C.AnswerTextArea value={answer} onChange={handleAnswerChange} placeholder="답변을 입력해주세요." />
+            {isEditing ? (
+              <C.SubmitButton onClick={handleAnswerUpdate}>수정</C.SubmitButton>
+            ) : (
+              <C.SubmitButton onClick={handleAnswerSubmit}>제출</C.SubmitButton>
+            )}
+          </C.AnswerSection>
+        );
+      }
     } else {
-      return <C.NoAnswerText>답변을 달고 있는 중이에요</C.NoAnswerText>;
+      return detailBoardData.answer ? (
+        <C.AnswerContainer>{detailBoardData.answer}</C.AnswerContainer>
+      ) : (
+        <C.NoAnswerText>답변을 달고 있는 중이에요</C.NoAnswerText>
+      );
     }
   };
 
@@ -128,9 +248,9 @@ function ContactBoardDetail() {
           <tbody>
             <C.TableRow>
               <C.TableHeader>제목</C.TableHeader>
-              <C.TableCell>{detailBoardData.title}</C.TableCell>
+              <C.TitleTableCell>{detailBoardData.title}</C.TitleTableCell>
               <C.TableHeader>등록일</C.TableHeader>
-              <C.CenterTableCell>{dayjs(detailBoardData.date).format('YYYY-MM-DD HH:mm')}</C.CenterTableCell>
+              <C.DateTableCell>{dayjs(detailBoardData.date).format('YYYY-MM-DD HH:mm')}</C.DateTableCell>
             </C.TableRow>
             <C.ContentTableRow>
               <C.TableHeader>내용</C.TableHeader>
@@ -138,17 +258,25 @@ function ContactBoardDetail() {
             </C.ContentTableRow>
           </tbody>
         </C.DetailTable>
-        {isAdmin ? (
-        <C.AnswerTable>
+        <C.BottomTable>
           <tbody>
-            <C.TableRow>
-              <C.AnswerTitle>AIT 답변</C.AnswerTitle>
-              <C.AnswerTableCell colSpan={1}>
-                {renderAnswerSection()}
-              </C.AnswerTableCell>
-            </C.TableRow>
+            <C.BottomRow>
+              <C.BottomStatusCell>{detailBoardData.answer ? '완료' : '처리중'}</C.BottomStatusCell>
+              <C.BottomCommentsCell onClick={toggleAnswerSection}>답변</C.BottomCommentsCell>
+            </C.BottomRow>
           </tbody>
-        </C.AnswerTable>):(<div>{isAdmin ? "true" : "false"}</div>)}
+        </C.BottomTable>
+        {showAnswer && (
+          <C.AnswerTable>
+            <tbody>
+              <C.RowTable>
+                <C.AnswerTableCell>
+                  {renderAnswerSection()}
+                </C.AnswerTableCell>
+              </C.RowTable>
+            </tbody>
+          </C.AnswerTable>
+        )}
       </C.SearchContainer>
     </C.PageContainer>
   );
